packages/mcp-server/pyproject.toml
[project]
name = "security-weekly-mcp-server"
version = "0.1.0"
description = "MCP Server for security weekly reports"
requires-python = ">=3.11"

dependencies = [
    "mcp[cli]>=1.2.0",
    "httpx[http2]>=0.25.0",
    "feedparser>=6.0",
    "pydantic>=2.5",
    "pyyaml>=6.0",
    "security-glossary-tw",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.3",
]

[tool.uv.sources]
security-glossary-tw = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/security_weekly_mcp"]
